+++ app/SessionsProcessor/GridModification/ModifyDurationMoment.php
@@ -43,7 +43,7 @@ class ModifyDurationMoment extends SessionMoment
     {
         /** @var Grid $grid */
         $grid                = (new Grid)->find($process->get(InitializationMoment::PROCESS_GRID_ID));
-        $grid->grid_duration = DurationMoment::parseDuration((int) $input);
+        $grid->grid_duration = DurationMoment::parseDuration((float) $input);
         $grid->save();
 
         $gridAdapter = GridAdapter::fromModel($grid);
